+++ resources/add_to_list.rb
@@ -1,3 +1,20 @@
+#
+# Copyright 2017, Sous Chefs
+#
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+# http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+#
+
 property :backup, [true, false], default: false
 property :delim, Array
 property :entry, String
@@ -20,6 +37,7 @@ action :edit do
 
   # eol on last line
   new[-1] += eol unless new[-1].to_s.empty?
+  current[-1] += eol unless current[-1].to_s.empty?
 
   file new_resource.path do
     content new.join(eol)